+++ src/codelens/registerCodeLense.ts
@@ -1,12 +1,10 @@
 import { languages, Disposable, DocumentSelector } from 'vscode'
 import { EsbmcCodeLensProvider } from './codeLensProvider'
+import { SUPPORTED_LANGUAGES } from '../languages'
 
-const selector: DocumentSelector = [
-  { language: 'c', scheme: 'file' },
-  { language: 'cpp', scheme: 'file' },
-  { language: 'sol', scheme: 'file' },
-  { language: 'py', scheme: 'file' }
-]
+const selector: DocumentSelector = SUPPORTED_LANGUAGES.map(
+  language => ({ language: language.id, scheme: 'file' })
+)
 
 export function registerCodeLens (): Disposable[] {
   return [